+++ src/lib/components/state/DisconnectedState.tsx
@@ -1,3 +1,4 @@
+import type { InputProps } from "@chakra-ui/react";
 import { Flex, Text } from "@chakra-ui/react";
 
 import { ConnectWalletBtn } from "../button/ConnectWallet";
@@ -7,13 +8,15 @@ import { StateImage } from "./StateImage";
 interface DisconnectedStateProps {
   text: string;
   helperText?: string;
+  my?: InputProps["my"];
 }
 
 export const DisconnectedState = ({
   text,
   helperText,
+  my = 0,
 }: DisconnectedStateProps) => (
-  <Flex direction="column" alignItems="center" gap={4}>
+  <Flex direction="column" alignItems="center" gap={4} my={my}>
     <StateImage imageVariant="disconnected" />
     <Flex align="center" justify="center">
       <ConnectWalletBtn />