+++ src/lib/pages/blocks/components/BlocksTable.tsx
@@ -2,7 +2,7 @@ import { useMobile } from "lib/app-provider";
 import { Loading } from "lib/components/Loading";
 import { Pagination } from "lib/components/pagination";
 import { usePaginator } from "lib/components/pagination/usePaginator";
-import { EmptyState } from "lib/components/state";
+import { EmptyState, ErrorFetching } from "lib/components/state";
 import { MobileTableContainer, TableContainer } from "lib/components/table";
 import { useBlocks } from "lib/services/blockService";
 
@@ -39,17 +39,9 @@ export const BlocksTable = ({ isViewMore }: BlocksTableProps) => {
     onSuccess: ({ total }) => setTotalData(total),
   });
 
-  if (isLoading) return <Loading withBorder />;
-  if (error)
-    return (
-      <EmptyState
-        imageVariant="not-found"
-        message="There is an error during fetching recent blocks."
-        withBorder
-      />
-    );
-
-  if (!data)
+  if (isLoading) return <Loading />;
+  if (error) return <ErrorFetching dataName="blocks" />;
+  if (!data?.total)
     return (
       <EmptyState
         imageVariant="empty"