+++ src/lib/services/accountService.ts
@@ -7,12 +7,10 @@ import {
   CELATONE_QUERY_KEYS,
   useBaseApiRoute,
   useWasmConfig,
+  useGovConfig,
 } from "lib/app-provider";
-import {
-  getAccountIdByAddressQueryDocument,
-  getAccountTypeByAddressQueryDocument,
-} from "lib/query";
-import type { AccountType, Addr, Nullable, Option } from "lib/types";
+import { getAccountTypeByAddressQueryDocument } from "lib/query";
+import type { AccountType, Addr, Option } from "lib/types";
 
 import {
   getAccountInfo,
@@ -20,32 +18,6 @@ import {
   getAccountTableCounts,
   type AccountTableCounts,
 } from "./account";
-
-export const useAccountId = (
-  walletAddress: Option<Addr>
-): UseQueryResult<Nullable<number>> => {
-  const { indexerGraphClient } = useCelatoneApp();
-
-  const queryFn = useCallback(async () => {
-    if (!walletAddress)
-      throw new Error("Error fetching account id: failed to retrieve address.");
-    return indexerGraphClient
-      .request(getAccountIdByAddressQueryDocument, { address: walletAddress })
-      .then<Nullable<number>>(
-        ({ accounts_by_pk }) => accounts_by_pk?.id ?? null
-      );
-  }, [indexerGraphClient, walletAddress]);
-
-  return useQuery(
-    [CELATONE_QUERY_KEYS.ACCOUNT_ID, indexerGraphClient, walletAddress],
-    queryFn,
-    {
-      enabled: Boolean(walletAddress),
-      retry: 1,
-      refetchOnWindowFocus: false,
-    }
-  );
-};
 
 export const useAccountType = (
   walletAddress: Option<Addr>,
@@ -96,11 +68,12 @@ export const useAccountTableCounts = (
   address: Addr
 ): UseQueryResult<AccountTableCounts> => {
   const endpoint = useBaseApiRoute("accounts");
+  const { enabled: isGov } = useGovConfig({ shouldRedirect: false });
   const { enabled: isWasm } = useWasmConfig({ shouldRedirect: false });
 
   return useQuery(
-    [CELATONE_QUERY_KEYS.TABLE_COUNTS, endpoint, address, isWasm],
-    async () => getAccountTableCounts(endpoint, address, isWasm),
+    [CELATONE_QUERY_KEYS.TABLE_COUNTS, endpoint, address, isGov, isWasm],
+    async () => getAccountTableCounts(endpoint, address, isGov, isWasm),
     { enabled: !!address, retry: 1, refetchOnWindowFocus: false }
   );
 };